linux/requirements.txt
ntplib==0.3.3
pyserial==2.7
